a copy of Oliver Jahn's MITgcm python package
This package comes with MITgcm code and resides in the folder MITgcm/utils/python/MITgcmutils. I copied here to make some minor changes. The following if from Oliver:
Current version of the file is in cvs under

MITgcm/utils/python/MITgcmutils

You need to have the directory

MITgcm/utils/python/MITgcmutils

on your python path (not its subdirectory of the same name!) for it to work.  One way of achieving this is to run

python setup.py --user

(or --prefix PREFIX instead of user if you have a special place for installing packages), but this is not necessary.  You can just leave the directory wherever it is and include it in your PYTHONPATH environment variable (or in sys.path in python).

Then you can do

from MITgcmutils import mds